feat: refactors hermit.js and tests to ts
